+++ creme/dummy.py
@@ -64,9 +64,6 @@ class NoChangeClassifier(base.MultiClassifier):
         probas[self.last_class] = 1
         return probas
 
-    def _more_tags(self):
-        return {'poor_score': True}
-
 
 class PriorClassifier(base.MultiClassifier):
     """Dummy classifier which uses the prior distribution.
@@ -118,9 +115,6 @@ class PriorClassifier(base.MultiClassifier):
     def predict_proba_one(self, x):
         return {label: count / self.n for label, count in self.counts.items()}
 
-    def _more_tags(self):
-        return {'poor_score': True}
-
 
 class StatisticRegressor(base.Regressor):
     """Dummy regressor that uses a univariate statistic to make predictions.
@@ -163,6 +157,3 @@ class StatisticRegressor(base.Regressor):
 
     def predict_one(self, x):
         return self.statistic.get()
-
-    def _more_tags(self):
-        return {'poor_score': True}