+++ creme/linear_model/__init__.py
@@ -5,6 +5,7 @@ from .fm import FMClassifier
 from .fm import FMRegressor
 from .glm import LinearRegression
 from .glm import LogisticRegression
+from .glm import PoissonRegression
 from .pa import PAClassifier
 from .pa import PARegressor
 from .softmax import SoftmaxRegression
@@ -19,5 +20,6 @@ __all__ = [
     'LogisticRegression',
     'PAClassifier',
     'PARegressor',
+    'PoissonRegression',
     'SoftmaxRegression'
 ]